+++ Sandbox/.gitignore
@@ -63,3 +63,4 @@ yarn-error.log
 .metro-health-check*
 
 android/app/google-services.json
+ios/GoogleService-Info.plist
