+++ Sandbox/ios/Sandbox/AppDelegate.mm
@@ -2,6 +2,8 @@
 
 #import <React/RCTBundleURLProvider.h>
 
+#import <Firebase.h>
+
 @implementation AppDelegate
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
@@ -11,6 +13,8 @@
   // They will be passed down to the ViewController used by React Native.
   self.initialProps = @{};
 
+  [FIRApp configure];
+  
   return [super application:application didFinishLaunchingWithOptions:launchOptions];
 }
